+++ CHANGES.md
@@ -2,6 +2,14 @@
 
 This document lists all changes of `python_base_app` with the most recent changes at the top.
 
+## Version 0.2.0 (July 18th, 2020)
+
+*   Add classes BaseUserHandler and UnixUserHandler
+*   Optional deployment of an apparmor configuration file
+*   Optional deployment of a tmpfiles.d configuration file
+*   Support for template in Debian configuration 
+*   Remove old Debian file tree before build package (to remove obsolete files)
+
 ## Version 0.1.11 (June 7th, 2020)
 
 *   Add Spanish localization (locale "es")